requirements.txt

fastapi
uvicorn
anthropic
sentence-transformers
faiss-cpu
python-dotenv
